requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.3
dj-rest-auth==7.0.1
Django==4.2.20
django-allauth==65.7.0
django-debug-toolbar==5.2.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
kombu==5.5.3
paystackapi==2.1.3
pillow==11.2.1
prompt_toolkit==3.0.51
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
redis==6.0.0
reportlab==4.4.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
